Пофиксил некорректную компановку компонентов Modal
