--- src/components/Modal/ModalBody/index.ts
@@ -1,1 +1,0 @@
-export * from './ModalBody';
